pyproject.toml
[tool.poetry]
name = "img2ico"
version = "0.1.0"
description = ""
authors = ["Eddie Lin <eddielin0926@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
dev = "img2ico.__main__:main"

[tool.poetry.dependencies]
python = "^3.9,<3.12"
pyside6 = "^6.4.2"
pyside6-fluent-widgets = "^1.3.3"
opencv-python = "^4.8.1.78"
pillow = "^10.1.0"

[tool.poetry.group.dev.dependencies]
pyinstaller = { git = "git@github.com:pyinstaller/pyinstaller.git" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
